git-svn-id: svn://svn.code.sf.net/p/partmodproject/code/trunk@18 3d63fe75-a90b-4577-b595-b0b07fb0d485

+++ Partmod/chs.hpp
@@ -22,11 +22,11 @@ public:
   CHS(long long lba);
   CHS();
 
-  bool SetCylinder(unsigned int);
+  bool SetCylinder(long long);
   bool SetHead(unsigned int);
   bool SetSector(unsigned int);
 
-  unsigned int GetCylinder();
+  unsigned long GetCylinder();
   unsigned int GetHead();
   unsigned int GetSector();
